refactor(app.js): move/extract methods from app.js to init.js; export new methods
